there has been some changes in common-service and application.yml files
changes:
application.yml -> defaultZone port for eureka was changed to 8761
Gson-> Gson is added as bean as named GsonConfig
AppUtil-> the return type from "gson.from" in method sendRequest was changed to APIResnpose, AppUtil is added as bean.
CommonService-> USER_SERVER_URL was wrong and fixed, method loadUserByUsername was missing "@Override" and fixed, webSecurity is excluded for this service.
Entity-> UserDto was missing annotation and fixed